requirements/base.txt
hydra-core==1.3.2
omegaconf==2.3.0
tqdm
wandb

numpy
scipy
pandas
matplotlib
opencv-python

gymnasium==1.2.0
dm-control==1.0.28
dm-env==1.6
dm-tree==0.1.9
imageio
imageio-ffmpeg
cloudpickle
h5py
lxml

ipython
ipykernel
pytest